+++ section/product/product.js
@@ -68,12 +68,34 @@ function populateTable(data) {
             deleteButton.style.backgroundColor = 'red';
             deleteButton.textContent = 'Delete';
             deleteButton.addEventListener('click', function() {
-                deleteProduct(item.id); // Pass the user id to the delete function
+                // SweetAlert2 confirmation dialog
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "Do you really want to delete this user?",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        // Call the deleteUser function only if confirmed
+                        deleteProduct(item.id); // Pass the user id to the delete function
+        
+                        // Optional: Show success message
+                        Swal.fire(
+                            'Deleted!',
+                            'The Product has been deleted.',
+                            'success'
+                        );
+                    }
+                });
             });
             deleteCell.appendChild(deleteButton);
         }
     });
 }
+
 
 function editProduct(user) {
     localStorage.setItem('userData', JSON.stringify(user));
